+++ tests/test_controller.py
@@ -1,4 +1,5 @@
 import pytest
+import re
 from httpcore import Request, Response
 
 from hishel._controller import (
@@ -58,9 +59,26 @@ def test_is_cachable_for_unsupported_method():
 def test_controller_with_unsupported_method():
     with pytest.raises(
         RuntimeError,
-        match="Hishel does not support the HTTP method `DELETE`. Please use either `GET` or `HEAD`.",
+        match="RFC9111 considers DELETE methods to be unsafe for caching.",
     ):
         Controller(cacheable_methods=["DELETE"])
+        
+def test_controller_with_invalid_method():
+    with pytest.raises(
+        RuntimeError,
+        match=re.escape("FISH is/are not valid HTTP method(s)."),
+    ):
+        Controller(cacheable_methods=["FISH",], allow_unsafe_methods=True)
+        
+def test_controller_with_allow_unsafe_methods():
+    controller = Controller(cacheable_methods=["DELETE"], allow_unsafe_methods=True)
+    
+    request = Request(b"DELETE", b"https://example.com", headers=[])
+    
+    response = Response(200, headers=[(b"Expires", b"some-date")])
+    
+    assert controller.is_cachable(request=request, response=response)
+    
 
 
 def test_is_cachable_for_unsupported_status():